[ContextGrids] Rename to ArtworkContextGrids and ctaDestination -> ctaHref
